index is restructured to take advantage of scala modifications with geolocation element not used gone, and a few elements moved around.
StoreSelect no longer is a snippet but comet actor, which listens in to Store messages from an actor to update the text input.

The ProductExchange actor is now a StoreProductExchange actor. This may change again but proves point that we can funnel different objects through same object. It does encode a last msg id to figure out which one to send out. Probably a heterogenous container should remove this ugliness.

GetStores now uses a Store as state as opposed to Box and used Actor publishing differently with actually a Store object.

ConfirmationReceiver is now more specific in XML transformation.